+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## UNRELEASED
 
-_Nothing yet_
+**Fixed:**
+- bpk-component-banner-alert
+  - Added support for a custom className and other arbitrary props
 
 ## 2017-08-08 - Enhancements to Chip, bug fixed for popover and modal